requirements.txt
langchain
streamlit
PyPDF2
transformers
python-dotenv